Fix up some comments/names

+++ src/mat91lib/sam4s/pit/pit.h
@@ -17,7 +17,7 @@ extern "C" {
 #define PIT_CLOCK_DIVISOR 1000
 
 /** Rate in Hz that the pit is incremented.  */
-#define PIT_RATE (F_CPU / PIT_CLOCK_DIVISOR)
+#define PIT_RATE (uint32_t)(F_CPU / PIT_CLOCK_DIVISOR)
 
 
 /** Define pit ticks.  */
